requirements.txt
pandas>=2.1
requests>=2.31
python-dotenv>=1.0
streamlit>=1.34
tqdm>=4.66
plotly>=5.0
